+++ Sources/OCCTBridge/src/OCCTBridge_Geom2d.mm
@@ -120,7 +120,7 @@ int32_t OCCTCurve2DEvaluateGrid(OCCTCurve2DRef curve,
         }
 
         NCollection_Array1<gp_Pnt2d> results = evaluator.EvaluateGrid(paramArr);
-        int32_t n = results.Size();
+        int32_t n = static_cast<int32_t>(results.Size());
         for (int32_t i = 0; i < n; i++) {
             const gp_Pnt2d& pt = results.Value(i + 1);
             outXY[i*2]   = pt.X();
@@ -145,7 +145,7 @@ int32_t OCCTCurve2DEvaluateGridD1(OCCTCurve2DRef curve,
         }
 
         NCollection_Array1<Geom2dGridEval::CurveD1> results = evaluator.EvaluateGridD1(paramArr);
-        int32_t n = results.Size();
+        int32_t n = static_cast<int32_t>(results.Size());
         for (int32_t i = 0; i < n; i++) {
             const Geom2dGridEval::CurveD1& r = results.Value(i + 1);
             outXY[i*2]     = r.Point.X();
